+++ eodag/plugins/download/aws.py
@@ -55,17 +55,22 @@ from eodag.utils.exceptions import (
     AuthenticationError,
     DownloadError,
     MisconfiguredError,
-    NoMatchingCollection,
     NotAvailableError,
     TimeOutError,
 )
-from eodag.utils.s3 import S3FileInfo, open_s3_zipped_object, stream_download_from_s3
+from eodag.utils.s3 import (
+    S3FileInfo,
+    list_files_in_s3_prefix,
+    open_s3_zipped_object,
+    stream_download_from_s3,
+)
 
 from .base import Download
 
 if TYPE_CHECKING:
     from mypy_boto3_s3 import S3ServiceResource
     from mypy_boto3_s3.client import S3Client
+    from mypy_boto3_s3.service_resource import BucketObjectsCollection, ObjectSummary
 
     from eodag.api.product import EOProduct
     from eodag.config import PluginConfig
@@ -256,7 +261,6 @@ class AwsDownload(Download):
                         file or with environment variables.
         :returns: The absolute path to the downloaded product in the local filesystem
         """
-
         if progress_callback is None:
             logger.info(
                 "Progress bar unavailable, please call product.download() instead of plugin.download()"
@@ -430,9 +434,7 @@ class AwsDownload(Download):
         progress_callback: ProgressCallback,
         executor: ThreadPoolExecutor,
     ):
-        """
-        Download file in zip from a prefix like `foo/bar.zip!file.txt`
-        """
+        """Download file in zip from a prefix like `foo/bar.zip!file.txt`."""
         if downloader_auth.s3_resource is None:
             logger.debug("Cannot check files in s3 zip without s3 resource")
             return bucket_names_and_prefixes
@@ -504,9 +506,9 @@ class AwsDownload(Download):
         progress_callback: ProgressCallback,
         **kwargs: Unpack[DownloadConf],
     ) -> tuple[Optional[str], Optional[str]]:
-        """
-        Preparation for the download:
+        """Prepare the download.
 
+        It means:
         - check if file was already downloaded
         - get file path
         - create directories
@@ -533,8 +535,7 @@ class AwsDownload(Download):
         return product_local_path, record_filename
 
     def _configure_safe_build(self, build_safe: bool, product: EOProduct):
-        """
-        Updates the product properties with fetch metadata if safe build is enabled
+        """Update the product properties with fetch metadata if safe build is enabled.
 
         :param build_safe: if safe build is enabled
         :param product: product to be updated
@@ -579,8 +580,7 @@ class AwsDownload(Download):
         ignore_assets: bool,
         complementary_url_keys: list[str],
     ) -> list[tuple[str, Optional[str]]]:
-        """
-        Retrieves the bucket names and path prefixes for the assets
+        """Retrieve the bucket names and path prefixes for the assets.
 
         :param product: product for which the assets shall be downloaded
         :param asset_filter: text for which the assets should be filtered
@@ -635,14 +635,13 @@ class AwsDownload(Download):
     def _get_unique_products(
         self,
         bucket_names_and_prefixes: list[tuple[str, Optional[str]]],
-        authenticated_objects: dict[str, Any],
+        authenticated_objects: dict[str, BucketObjectsCollection],
         asset_filter: Optional[str],
         ignore_assets: bool,
         product: EOProduct,
         raise_error: bool = True,
-    ) -> set[Any]:
-        """
-        Retrieve unique product chunks based on authenticated objects and asset filters
+    ) -> set[ObjectSummary]:
+        """Retrieve unique product chunks based on authenticated objects and asset filters.
 
         :param bucket_names_and_prefixes: list of bucket names and corresponding path prefixes
         :param authenticated_objects: available objects per bucket
@@ -652,12 +651,15 @@ class AwsDownload(Download):
         :param raise_error: raise error if there is nothing to download
         :return: set of product chunks that can be downloaded
         """
-        product_chunks: list[Any] = []
+        product_chunks: list[ObjectSummary] = []
         for bucket_name, prefix in bucket_names_and_prefixes:
             # unauthenticated items filtered out
             if bucket_name in authenticated_objects.keys():
                 product_chunks.extend(
-                    authenticated_objects[bucket_name].filter(Prefix=prefix)
+                    list_files_in_s3_prefix(
+                        prefix or "",
+                        authenticated_objects[bucket_name],
+                    )
                 )
 
         unique_product_chunks = set(product_chunks)
@@ -672,12 +674,15 @@ class AwsDownload(Download):
                 )
             )
             if not unique_product_chunks and raise_error:
-                raise NotAvailableError(
-                    rf"No file basename matching re.fullmatch(r'{asset_filter}') was found in {product.remote_location}"
+                msg = (
+                    rf"No file basename matching re.fullmatch(r'{asset_filter}') was found in "
+                    f"{product.remote_location}"
                 )
+                raise NotAvailableError(msg)
 
         if not unique_product_chunks and raise_error:
-            raise NoMatchingCollection("No product found to download.")
+            msg = f"No downloadable files found for {product}."
+            raise NotAvailableError(msg)
 
         return unique_product_chunks
 
@@ -691,8 +696,7 @@ class AwsDownload(Download):
         timeout: float = DEFAULT_DOWNLOAD_TIMEOUT,
         **kwargs: Unpack[DownloadConf],
     ) -> StreamResponse:
-        """
-        Stream EO product data as a FastAPI-compatible `StreamResponse`, with support for partial downloads,
+        """Stream EO product data as a FastAPI-compatible `StreamResponse`, with support for partial downloads,
         asset filtering, and on-the-fly compression.
 
         This method streams data from one or more S3 objects that belong to a given EO product.
@@ -855,7 +859,7 @@ class AwsDownload(Download):
     def get_product_bucket_name_and_prefix(
         self, product: EOProduct, url: Optional[str] = None
     ) -> tuple[str, Optional[str]]:
-        """Extract bucket name and prefix from product URL
+        """Extract bucket name and prefix from product URL.
 
         :param product: The EO product to download
         :param url: (optional) URL to use as product.location
@@ -880,7 +884,7 @@ class AwsDownload(Download):
         return bucket, prefix
 
     def check_manifest_file_list(self, product_path: str) -> None:
-        """Checks if products listed in manifest.safe exist"""
+        """Check if products listed in manifest.safe exist."""
         manifest_path_list = [
             os.path.join(d, x)
             for d, _, f in os.walk(product_path)
@@ -904,7 +908,7 @@ class AwsDownload(Download):
                 logger.warning("SAFE build: %s is missing" % safe_file.get("href"))
 
     def finalize_s2_safe_product(self, product_path: str) -> None:
-        """Add missing dirs to downloaded product"""
+        """Add missing dirs to downloaded product."""
         try:
             logger.debug("Finalize SAFE product")
             manifest_path_list = [
@@ -970,7 +974,7 @@ class AwsDownload(Download):
         dir_prefix: Optional[str] = None,
         build_safe: bool = False,
     ) -> str:
-        """Get chunk SAFE destination path"""
+        """Get chunk SAFE destination path."""
         if not build_safe:
             if dir_prefix is None:
                 dir_prefix = chunk.key